+++ src/client/AM.PA.MacOs/PersonalAnalytics/Trackers/EmotionTracker/EmotionTrackerQueries.swift
@@ -45,44 +45,4 @@ class EmotionTrackerQueries {
             print(error)
         }
     }
-    
-    
-    //TODO: refactor
-    struct EmotionalStateEntry {
-        var timestamp: Double
-        var activity: String
-        var valence: Int
-        var arousal: Int
-    }
-
-    static func fetchEmotionalStateSince(time: Double) -> [EmotionalStateEntry] {
-        let dbController = DatabaseController.getDatabaseController()
-        var results: [EmotionalStateEntry] = []
-        let timeStr = DateFormatConverter.interval1970ToDateStr(interval: time)
-        
-        do {
-            let query = """
-                        SELECT * FROM emotional_state
-                        WHERE timestamp >= '\(timeStr)'
-                        ORDER BY timestamp
-                        """
-            
-            let rows = try dbController.executeFetchAll(query: query)
-            
-            for row in rows {
-
-                let timestamp: Double = DateFormatConverter.dateStrToInterval1970(str: row["timestamp"])
-                let activity: String = row["activity"]
-                let valence: Int = row["valence"]
-                let arousal: Int = row["arousal"]
-
-                results.append(EmotionalStateEntry(timestamp: timestamp, activity: activity, valence: valence, arousal: arousal))
-            }
-
-        } catch {
-            print(error)
-        }
-
-        return results
-    }
 }
